consolidate components
