--- sales-coach-avatar/src/app/_views/demo/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./LeftPanel";
-export * from "./RightPanel";
