--- sales-coach-avatar/src/app/_views/demo/index.ts
@@ -1,1 +1,0 @@
-export * from "./demo";
