--- sales-coach-avatar/src/app/_views/description/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./LeftPanel";
-export * from "./RightPanel";
