--- sales-coach-avatar/src/app/_views/description/index.ts
@@ -1,1 +1,0 @@
-export * from "./description";
