--- sales-coach-avatar/src/app/_views/initial/index.ts
@@ -1,1 +1,0 @@
-export * from "./initial";
